# Conflicts:
#	src/main/java/com/swp/autocarwash/booking/repository/BookingRepository.java
#	src/main/java/com/swp/autocarwash/booking/service/impl/BookingServiceImpl.java
#	src/main/java/com/swp/autocarwash/common/exception/code/ErrorCode.java
#	src/main/java/com/swp/autocarwash/common/exception/handler/GlobalExceptionHandler.java
#	src/main/java/com/swp/autocarwash/subscription/repository/FamilySubscriptionRepository.java
#	src/main/java/com/swp/autocarwash/subscription/repository/UnlimitSubscriptionRepository.java

+++ src/main/java/com/swp/autocarwash/common/contract/promotion/VoucherContract.java
@@ -11,14 +11,14 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @NoArgsConstructor
 public class VoucherContract {
-    private Integer id;
+    private Long id;
     private String voucherCode;
     private Integer discountPercentage;
     private BigDecimal minOrderValue;
     private boolean valid;
 
 
-    public VoucherContract(Integer id, String voucherCode, Integer discountPercentage, BigDecimal minOrderValue) {
+    public VoucherContract(Long id, String voucherCode, Integer discountPercentage, BigDecimal minOrderValue) {
         this.id = id;
         this.voucherCode = voucherCode;
         this.discountPercentage = discountPercentage;
